comparators mostly working. finally

+++ crates/backend/src/fpga/mod.rs
@@ -57,7 +57,7 @@ impl JITBackend for FPGABackend {
         plot: String,
         name: String,
         config: Option<DeviceConfig>,
-        _options: &CompilerOptions,
+        options: &CompilerOptions,
     ) {
         let path = format!("{}/{}", plot, name);
         self.link.name = name;
@@ -79,8 +79,10 @@ impl JITBackend for FPGABackend {
         println!("compile");
         self.config.compile(Path::new(&format!("FPGA/bin/{}/prj", self.path)));
         println!("done");
-        _ = copy(Path::new(&format!("FPGA/bin/{}/prj/RoC.sof", self.path)), Path::new(&format!("FPGA/bin/{}/RoC.sof", self.path)));
-        _ = remove_dir_all(Path::new(&format!("FPGA/bin/{}/prj", self.path)));  
+        if !options.compile_verilog {
+            _ = copy(Path::new(&format!("FPGA/bin/{}/prj/RoC.sof", self.path)), Path::new(&format!("FPGA/bin/{}/RoC.sof", self.path)));
+            _ = remove_dir_all(Path::new(&format!("FPGA/bin/{}/prj", self.path)));  
+        }
     }
 
     fn run(&mut self) {
